# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponseFormat": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"error": {
						"type": "string"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegisterAdminBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPBanBody": {
				"properties": {
					"ipAddress": {
						"type": "string"
					}
				},
				"required": [
					"ipAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateApiLimitBody": {
				"properties": {
					"youtubeLimit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"youtubeLimit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCacheSizeBody": {
				"properties": {
					"searchCacheLineSize": {
						"type": "number",
						"format": "double"
					},
					"metadataCacheLineSize": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"userGroupAuth": {
				"type": "http",
				"scheme": "bearer"
			},
			"adminGroupAuth": {
				"type": "http",
				"scheme": "bearer"
			},
			"adminAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "Jukebox API",
		"version": "1.0.0",
		"description": "Create a group with friends, add your music, vote in real time to choose the next song, and listen to it together!",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "CapyBlaze",
			"email": "capyblaze@gmail.com"
		}
	},
	"paths": {
...
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "User logged in",
											"data": {
												"token": "1c565e42-3f76-4f78-b9a4-26dcddac786c",
												"name": "admin"
											},
											"timestamp": "2026-06-29T16:14:18.099Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Authenticate as an administrator and receive the bearer token required for admin routes.",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterAdminBody"
							}
						}
					}
				}
			}
		},
		"/admin/group/{groupId}": {
			"delete": {
				"operationId": "DeleteGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Group deleted",
											"timestamp": "2026-06-29T16:18:49.137Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"404": {
						"description": "Group not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Delete a group.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "groupId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/user/{userId}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "User deleted",
											"timestamp": "2026-06-29T16:18:49.137Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Delete a user.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/ip/ban": {
			"post": {
				"operationId": "BanIP",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "IP banned",
											"data": {
												"ipAddress": "203.0.113.42"
											},
											"timestamp": "2026-06-29T16:18:49.137Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing IP address",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Ban an IP address from accessing the API.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPBanBody"
							}
						}
					}
				}
			}
		},
		"/admin/ip/unban": {
			"post": {
				"operationId": "UnbanIP",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "IP unbanned",
											"data": {
												"ipAddress": "203.0.113.42"
											},
											"timestamp": "2026-06-29T16:21:02.254Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing IP address",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Remove a ban from an IP address.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPBanBody"
							}
						}
					}
				}
			}
		},
		"/admin/ip/banned": {
			"get": {
				"operationId": "GetBannedIPs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Banned IPs fetched",
											"data": [
												{
													"ipAddress": "203.0.113.42",
													"bannedAt": "2026-06-29T16:18:49.126Z"
												}
											],
											"timestamp": "2026-06-29T16:20:28.338Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get a list of all banned IP addresses.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/admin/api/limit": {
			"get": {
				"operationId": "GetApiLimit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "YouTube API limit fetched",
											"data": {
												"youtubeLimit": 10000
											},
											"timestamp": "2026-06-29T16:32:05.221Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get the current daily YouTube API call limit.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateApiLimit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "YouTube API limit updated",
											"data": {
												"youtubeLimit": 10000
											},
											"timestamp": "2026-06-29T16:23:09.664Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Update the daily YouTube API call limit.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateApiLimitBody"
							}
						}
					}
				}
			}
		},
		"/admin/api/usage": {
			"get": {
				"operationId": "GetApiUsage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "YouTube API usage fetched",
											"data": [
												{
													"id": "cmqzdv4130000xkkli9ocs5y6",
													"provider": "youtube",
													"startDate": "2026-06-29T00:00:00.000Z",
													"endDate": "2026-06-29T23:59:59.999Z",
													"used": 101,
													"lastUsed": "2026-06-29T15:39:13.575Z"
												}
											],
											"timestamp": "2026-06-29T16:32:42.506Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get the current YouTube API usage for the day.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/admin/cache/size": {
			"get": {
				"operationId": "GetCacheSize",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Cache sizes fetched",
											"data": {
												"searchCacheLineSize": 100,
												"metadataCacheLineSize": 100
											},
											"timestamp": "2026-06-29T23:09:59.432Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get the current maximum cache line sizes for search and metadata caches.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateCacheSize",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Cache sizes updated",
											"data": {
												"searchCacheLineSize": 200,
												"metadataCacheLineSize": 500
											},
											"timestamp": "2026-06-29T16:24:06.726Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Update the maximum number of cache lines for the search and metadata caches.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCacheSizeBody"
							}
						}
					}
				}
			}
		},
		"/admin/cache/clear": {
			"delete": {
				"operationId": "ClearCache",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Cache cleared",
											"timestamp": "2026-06-29T16:24:30.768Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Clear all server-side caches (search and metadata).",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/admin/cache/cleanup": {
			"post": {
				"operationId": "CleanupCache",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Cache cleanup completed",
											"timestamp": "2026-06-29T23:10:45.888Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Cleanup the search and metadata caches.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		},
		"/admin/stats": {
			"get": {
				"operationId": "GetStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Admin stats fetched",
											"data": {
												"groups": 2,
												"users": 3,
												"apiVersion": "v1",
												"uptime": 22045,
												"memory": {
													"usedMB": 18570,
													"freeMB": 13555,
													"totalMB": 32125,
													"usagePercent": 57.81
												},
												"cpuUsage": "9.36%"
											},
											"timestamp": "2026-06-29T16:25:03.147Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get global game totals and host resource statistics for the admin dashboard.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"adminAuth": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "https://jukebox.capyblaze.hackclub.app/"
		}
	]
}